Image loader and examples
I started work on an image and texture loader to try and remove code repitition. I also updated the example with comments and removed code bloat (commented out code, or code that I probably won't use again). I also added a new function in the ibl module that helps with generating and loading the necessary maps for pbr.