Improved Portability
The shader files are now included when compiling this lib. This fixes a severe bug where the library was useless unless you manually included the necessary files. I also added custom scene variable support to the RenderScene so when expanding on the library you can have custom scene variables for your shaders.